TOTS-80 Moved the browser storage hooks to a separate file.
